Further refactor common logic away into Flashbar hook
